- show message in AcceptActivity for ExerciseMode - trigger ExerciseMode after accepting

+++ core/interfaces/src/main/kotlin/app/aaps/core/interfaces/rx/weardata/EventData.kt
@@ -170,7 +170,7 @@ sealed class EventData : Event() {
     data class ActionWizardConfirmed(val timeStamp: Long) : EventData()
 
     @Serializable
-    data class ActionExerciseModeConfirmed(val percentage: Double = 0.0, val duration: Int = 0, val timeshift: Double = 0.0) : EventData()
+    data class ActionExerciseModeConfirmed(val percentage: Int = 0, val duration: Int = 0, val timeshift: Int = 0) : EventData()
 
     @Serializable
     data class ActionTempTargetConfirmed(val isMgdl: Boolean = true, val duration: Int = 0, val low: Double = 0.0, val high: Double = 0.0) : EventData()
